refactor: extract MapViewModel status filter tests to own file
Move status filtering tests to MapViewModelStatusFilterTests.swift to
keep MapViewModelTests.swift under the 400-line SwiftLint threshold.
